[IMP] sell_product_kit: report creation print order done

+++ sell_product_kit/models/product_kit_model.py
@@ -13,13 +13,12 @@ class ProductKit(models.TransientModel):
         defaults = super().default_get(fields_list)
         product_kit_line_ids=[]
         
-        breakpoint()
-        if(len(sale_order_line_id.linked_line_ids)>0):
-            for c_pro in sale_order_line_id.linked_line_ids:
+        if(len(sale_order_line_id.linked_product_kit_ids)>0):
+            for c_pro in sale_order_line_id.linked_product_kit_ids:
                 product_kit_line_ids.append({
                     'product_id':c_pro.product_id.id,
                     'quantity': c_pro.product_uom_qty,
-                    'price': 1
+                    'price': c_pro.prev_filled_unit_price
                 })
         else:
             for c_pro in product_template.sub_products_ids:
@@ -42,21 +41,26 @@ class ProductKit(models.TransientModel):
         
         sub_product_total= 0
         
-        for prod in template_sub_products:
-            self.env['sale.order.line'].create({
-                'linked_line_id': sale_order_line.id,
-                'order_id': sale_order_line.order_id.id,
-                'product_id': prod.product_id.id,
-                'product_uom_qty': prod.quantity,
-                'price_unit': 0,
-                'customer_lead':0
-            })
-            sub_product_total+= prod.quantity * prod.price
+        #is no sale.order.line exist related to the current product order line only then create else update
+        if(len(sale_order_line.linked_product_kit_ids)==0):    
+            for prod in template_sub_products:
+                self.env['sale.order.line'].create({
+                    'linked_product_kit_id': sale_order_line.id,
+                    'order_id': sale_order_line.order_id.id,
+                    'product_id': prod.product_id.id,
+                    'product_uom_qty': prod.quantity,
+                    'price_unit': 0,
+                    'customer_lead':0,
+                    'prev_filled_unit_price': prod.price
+                })
+                sub_product_total+= prod.quantity * prod.price
 
-        sale_order_line.update({
-            'price_subtotal': sale_order_line.price_subtotal+sub_product_total
-        })
-        template_sub_products.unlink()
+            sale_order_line.update({
+                'price_subtotal': sale_order_line.price_subtotal+sub_product_total
+            })
+            template_sub_products.unlink()
+        #TODO TO UPDATE THE LINE'S PRICE AND QUNATITY
+        
         return True
 
 class ProductKitLines(models.TransientModel):
